✏️ update : 그룹 삭제 api 연결 완료!
그룹 삭제 api 연결 완료!, 그룹 생성과 그룹 수정 api 수정 예정

> Related to : #38

+++ app/src/main/java/com/example/gajamap/data/model/RadiusData.kt
@@ -8,7 +8,7 @@ data class RadiusResponse(
 )
 data class RadiusClient(
     @SerializedName("clientId")
-    val clientId: Int,
+    val clientId: Long,
     @SerializedName("groupInfo")
     val groupInfo: GroupInfoData?,
     @SerializedName("clientName")
@@ -26,7 +26,7 @@ data class RadiusClient(
 )
 data class GroupInfoData(
     @SerializedName("groupId")
-    val groupId: Int,
+    val groupId: Long,
     @SerializedName("groupName")
     val groupName: String
 )
